build_bsp: Toggle all BUILD_@testsuite@ configuration settings
Late 2023, the behavior for BUILD_TESTS to override the individual
BUILD_ variables was changed. Each test suite needs to be enabled
in order to build all tests. This change showed up on the RTEMS
Build mailing list as test runs from the OAR Cron Sweep only building
and running samples.